collect all customer functions inside customerController.js
